% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusvisvarsellcm.R
\name{clusvisVarselclm}
\alias{clusvisVarselclm}
\title{This function performs the variable selection and the maximum likelihood estimation of the Latent Class Model}
\usage{
clusvisVarselclm(resvarsellcm, sample.size = 2000, maxit = 10^3,
  nbrandomInit = 4 * resvarsellcm@model@g, nbcpu = 3)
}
\arguments{
\item{maxit}{numeric. It limits the number of iterations for the Quasi-Newton algorithm (default 1000)}

\item{nbrandomInit}{numeric. It defines the number of random initialization of the Quasi-Newton algorithm.}

\item{logtik.estim}{matrix. It contains the probabilities of classification used for parameter inference.}

\item{prop}{vector. It contains the class proportions (by default, classes have same proportion).}

\item{logtik.obs}{matrix. It contains the probabilities of classification of the clustered sample. If missing, logtik.estim is used.}
}
\value{
Returns a list
}
\description{
This function performs the variable selection and the maximum likelihood estimation of the Latent Class Model
}
\examples{
set.seed(123)
}
