% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusvis.R
\docType{data}
\name{congress}
\alias{congress}
\title{Real categorical data set: Congressional Voting Records Data Set}
\description{
This data set includes votes for each of the U.S. House of Representatives Congressmen on the 16 key votes identified by the CQA. The CQA lists nine different types of votes: voted for, paired for, and announced for (these three simplified to yea), voted against, paired against, and announced against (these three simplified to nay), voted present, voted present to avoid conflict of interest, and did not vote or otherwise make a position known (these three simplified to an unknown disposition).
}
\examples{
  data(congress)
}
\references{
Congressional Quarterly Almanac, 98th Congress, 2nd session 1984, Volume XL: Congressional Quarterly Inc. Washington, D.C., 1985.

Schlimmer, J. C. (1987). Concept acquisition through representational adjustment. Doctoral dissertation, Department of Information and Computer Science, University of California, Irvine, CA.

Website: https://archive.ics.uci.edu/ml/datasets/congressional+voting+records
}
\keyword{datasets}
