% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDensityClusVisu}
\alias{plotDensityClusVisu}
\title{Function for visualizing the clustering results}
\usage{
plotDensityClusVisu(res, dim = c(1, 2), threshold = 0.95, add.obs = FALSE,
  positionlegend = "topright", xlim = NULL, ylim = NULL)
}
\arguments{
\item{res}{object return by function clusvis}

\item{dim}{numeric. This vector of size two choose the axes to represent.}

\item{threshold}{numeric. It contains the thersholds used for computing the level curves.}

\item{add.obs}{boolean. If TRUE, coordinnates of the observations are plotted.}

\item{add.centers}{boolean. If TRUE, centers are plotted.}

\item{col.dots}{numeric/character. It specifies the color of the dots.}

\item{col.centers}{numeric/character. It specifies the color of the centers.}

\item{cex.dots}{numeric. It specifies the size of the observations.}

\item{...}{other parameters of the plot function.}
}
\description{
Function for visualizing the clustering results
}
\examples{
set.seed(123)
}
